refactor:
QueueEntry 위치 변경
